:art: [#2174] changed map location of leaflet map component api
