+++ src/openforms/api/urls.py
@@ -114,6 +114,7 @@ urlpatterns = [
                     include("openforms.appointments.api.urls"),
                 ),
                 path("formio/", include("openforms.formio.api.urls")),
+                path("geo/", include("openforms.contrib.kadaster.api.urls")),
                 path("i18n/", include("openforms.translations.api.urls")),
                 path("", include(router.urls)),
                 path("", include(forms_router.urls)),
