+++ src/openforms/formio/api/serializers.py
@@ -51,19 +51,3 @@ class TemporaryFileUploadSerializer(serializers.Serializer):
             kwargs={"uuid": instance.uuid},
             request=request,
         )
-
-
-class MapSearchLatitudeLongitudeSerializer(serializers.Serializer):
-    lat = serializers.FloatField(label=_("Latitude"))
-    lng = serializers.FloatField(label=_("Longitude"))
-
-
-class MapSearchRijksDriehoekSerializer(serializers.Serializer):
-    x = serializers.FloatField(label=_("X"))
-    y = serializers.FloatField(label=_("Y"))
-
-
-class MapSearchSerializer(serializers.Serializer):
-    label = serializers.CharField(label=_("weergave naam"))
-    latLng = MapSearchLatitudeLongitudeSerializer()
-    rd = MapSearchRijksDriehoekSerializer()
